Add culling.

+++ src/Descriptors.cpp
@@ -196,6 +196,6 @@ namespace VulkanEngine
         {
             write.dstSet = set;
         }
-        vkUpdateDescriptorSets(pool.device.GetDevice(), writes.size(), writes.data(), 0, nullptr);
+        vkUpdateDescriptorSets(pool.device.GetDevice(), static_cast<uint32_t>(writes.size()), writes.data(), 0, nullptr);
     }
 }
